Way too many changes

+++ src/main/kotlin/com/hedley/groceriesstats/itempurchase/ItemPurchaseDTO.kt
@@ -3,4 +3,4 @@ package com.hedley.groceriesstats.itempurchase
 import com.hedley.groceriesstats.items.ItemDTO
 import com.hedley.groceriesstats.purchase.PurchaseDTO
 
-data class ItemPurchaseDTO(val item: ItemDTO, val purchase: PurchaseDTO, val value: Double)
+data class ItemPurchaseDTO(val item: ItemDTO, val purchase: PurchaseDTO, val value: Float)
